adding some validation

+++ controllers/Booking.js
@@ -27,7 +27,7 @@ router.get("/:bookingId", async (req, res) => {
     res.status(500).json(error);
   }
 });
-
+//update a booking
 router.put("/:bookingId", async (req, res) => {
   try {
     const Book = await Booking.findByIdAndUpdate(
@@ -40,9 +40,15 @@ router.put("/:bookingId", async (req, res) => {
     res.status(500).json(error);
   }
 });
-
+//delete a booking
 router.delete("/:bookingId", async (req, res) => {
   try {
+    const book = await Booking.findById(req.params.bookingId);
+
+    if (!book.userId.equals(req.user._id)) {
+      return res.status(403).send("You're not allowed to do that!");
+    }
+
     const Book = await Booking.findByIdAndDelete(req.params.bookingId);
     res.status(200).json(Book);
   } catch (error) {
